+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## [161.16.0](https://github.com/Sage/carbon/compare/v161.15.0...v161.16.0) (2026-08-06)
+
+### Features
+
+* **search:** add non en-GB locale support ([3ed3147](https://github.com/Sage/carbon/commit/3ed31478f3453a322ba2e88c17868b447e26f393))
+
+## [161.15.0](https://github.com/Sage/carbon/compare/v161.14.0...v161.15.0) (2026-08-06)
+
+### Features
+
+* **icon:** use color preset for color prop ([810f230](https://github.com/Sage/carbon/commit/810f230a93912da4436026d0576db5171c81c64d))
+
 ## [161.14.0](https://github.com/Sage/carbon/compare/v161.13.2...v161.14.0) (2026-08-06)
 
 ### Features